Проведен рефакторинг тестов метода copyDeep и equals
